added more functionality to users

+++ GameReviewsAPI/GameReviews.API/GameReviews.API/Controllers/AccountsController.cs
@@ -26,11 +26,12 @@ namespace GameReviews.API.Controllers
         private readonly ApplicationDbContext _context;
         private readonly IReviewsService _reviewsService;
         private readonly IMapper _mapper;
+        private readonly IEmailSender _emailSender;
         private string container = "userspictures";
         private string userType = "User";
-        private string pictureString = "https://gamereviewsapi.blob.core.windows.net/userspictures/23bc1d0e-6d9a-4b93-a5c4-cdc43be653f3.png";
+        private string pictureString = "https://gamereviewsapi.blob.core.windows.net/userspictures/d06e8f56-16ec-47b2-978a-03dcda797947.png";
 
-        public AccountsController(UserManager<ApplicationUser> userManager, SignInManager<ApplicationUser> signInManager,
+        public AccountsController(UserManager<ApplicationUser> userManager, SignInManager<ApplicationUser> signInManager, IEmailSender emailSender,
             IConfiguration configuration, IFileStorageService fileStorageService, ApplicationDbContext context, IMapper mapper, IReviewsService reviewsService)
         {
             _configuration = configuration;
@@ -40,6 +41,7 @@ namespace GameReviews.API.Controllers
             _context = context;
             _mapper = mapper;
             _reviewsService = reviewsService;
+            _emailSender = emailSender;
         }
         [HttpPost("create")]
         public async Task<ActionResult<AuthenticationResponseDTO>> Create([FromForm] UserCreationDTO userCreation)
@@ -47,7 +49,6 @@ namespace GameReviews.API.Controllers
             var user = new ApplicationUser { UserName = userCreation.Email, Email = userCreation.Email };
             user.Type = userType;
             var result = await _userManager.CreateAsync(user, userCreation.Password);
-
             var userCredentials = new UserCredentialsDTO() { Email = userCreation.Email, Password = userCreation.Password };
             if (result.Succeeded)
             {
@@ -61,7 +62,8 @@ namespace GameReviews.API.Controllers
                 }
                 await _userManager.UpdateAsync(user);
 
-                return BuildToken(userCredentials, user.ProfilePicture);
+                var claimToSend = new Claim("type", "User");
+                return BuildToken(user.Email, user.ProfilePicture, claimToSend);
             }
             else
             {
@@ -79,28 +81,109 @@ namespace GameReviews.API.Controllers
                 var user = await _userManager.FindByEmailAsync(userCredentials.Email);
                 var claims = await _userManager.GetClaimsAsync(user);
                 Claim claimToSend = new Claim("", "");
-                foreach (var claim in claims)
+                if (claims.Count() != 0)
                 {
-                    if (claim.Value == "admin")
+                    foreach (var claim in claims)
                     {
-                        claimToSend = new Claim("role", "admin");
+                        if (claim.Value == "admin")
+                        {
+                            claimToSend = new Claim("role", "admin");
+                        }
                     }
                 }
-                return BuildToken(userCredentials, user.ProfilePicture, claimToSend);
+                else
+                {
+                    var type = user.Type;
+                    claimToSend = new Claim("type", type);
+                }
+                return BuildToken(user.Email, user.ProfilePicture, claimToSend);
             }
             else
             {
                 return BadRequest("Incorrect Login");
             }
+        }
+
+        [HttpPost("changeCredentials")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+        public async Task<ActionResult<AuthenticationResponseDTO>> ChangeDetails([FromForm] UserCredentailsToChangeDTO userCredentials)
+        {
+            var result = await _signInManager.PasswordSignInAsync(userCredentials.OldEmail,
+                userCredentials.OldPassword, isPersistent: false, lockoutOnFailure: false);
+            if (result.Succeeded)
+            {
+                var user = await _userManager.FindByEmailAsync(userCredentials.OldEmail);
+                if (userCredentials.NewEmail != null)
+                {
+                    try
+                    {
+                        user.Email = userCredentials.NewEmail;
+                        user.UserName = userCredentials.NewEmail;
+                        var resultEmail = await _userManager.UpdateAsync(user);
+                    }
+                    catch (Exception e)
+                    {
+                        return BadRequest(e.Message);
+                    }
+                }
+                if (userCredentials.NewPassword != null)
+                {
+                    var resultPassword = await _userManager.ChangePasswordAsync(user, userCredentials.OldPassword, userCredentials.NewPassword);
+                    if (resultPassword.Succeeded == false)
+                    {
+                        return BadRequest(resultPassword.Errors);
+                    }
+                }
+                if (userCredentials.NewProfilePicture != null)
+                {
+                    try
+                    {
+                        if (user.ProfilePicture != pictureString)
+                        {
+                            await _fileStorageService.DeleteFile(user.ProfilePicture, container);
+                        }
+                        user.ProfilePicture = await _fileStorageService.SaveFile(container, userCredentials.NewProfilePicture);
+                        var resultPicture = await _userManager.UpdateAsync(user);
+                    }
+                    catch (Exception e)
+                    {
+                        return BadRequest(e.Message);
+                    }
+                }
+                var claims = await _userManager.GetClaimsAsync(user);
+                Claim claimToSend = new Claim("", "");
+                if (claims.Count() != 0)
+                {
+                    foreach (var claim in claims)
+                    {
+                        if (claim.Value == "admin")
+                        {
+                            claimToSend = new Claim("role", "admin");
+                        }
+                    }
+                }
+                else
+                {
+                    var type = user.Type;
+                    claimToSend = new Claim("type", type);
+                }
+                return BuildToken(user.Email, user.ProfilePicture, claimToSend);
+            }
+            else
+            {
+                return BadRequest("Incorrect Password");
+            }
 
         }
+
+
+
 
         [HttpGet("")]
         [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = "isAdmin")]
         public async Task<ActionResult<List<UserDTO>>> Get([FromQuery] PaginationDTO paginationDTO)
         {
             var queryable = _userManager.Users.AsQueryable();
-            //var queryable = _context.Users.AsQueryable();
             await HttpContext.InsertParametersPaginationInHeader(queryable);
             var users = await queryable.OrderBy(x => x.Email).Paginate(paginationDTO).ToListAsync();
             var result = _mapper.Map<List<UserDTO>>(users);
@@ -171,32 +254,41 @@ namespace GameReviews.API.Controllers
             }
             if (user is null) return NotFound();
             await _userManager.DeleteAsync(user);
-            if (user.ProfilePicture != null)
+            if (user.ProfilePicture != null && user.ProfilePicture != pictureString)
             {
+
                 await _fileStorageService.DeleteFile(user.ProfilePicture, container);
             }
             return Ok();
         }
 
-
-        private AuthenticationResponseDTO BuildToken(UserCredentialsDTO userCredentialsDTO, string? profilePicture, Claim? claim = null)
+        [HttpPost("sendEmail")]
+        public async Task<ActionResult> SendEmail([FromForm] SendEmailDTO sendEmailDTO)
         {
+            var admins = await _userManager.Users.Where(x => x.Type == "Admin").Select(x => x.Email).ToListAsync();
+            var response = await _emailSender.SendEmailAsync(sendEmailDTO.Email, sendEmailDTO.Text, admins, sendEmailDTO.Image);
+            return Ok(response);
+        }
 
+
+
+
+
+        private AuthenticationResponseDTO BuildToken(string email, string? profilePicture, Claim? claim = null)
+        {
             var claims = new List<Claim>()
             {
-                new Claim("email",userCredentialsDTO.Email),
-
+                new Claim("email",email),
             };
+
             claims.Add(claim);
             if (profilePicture != null)
             {
                 claims.Add(new Claim("profilePicture", profilePicture));
             }
-
             var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["jwtkey"]));
             var credentials = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
-            //var expiration = DateTime.UtcNow.AddYears(1);
-            var expiration = DateTime.UtcNow.AddHours(2);  ///aici de modificat daca nu vreau un an !!
+            var expiration = DateTime.UtcNow.AddHours(2);  ///aici de modificat timpul!!
             var token = new JwtSecurityToken(issuer: null, audience: null, claims: claims, expires: expiration, signingCredentials: credentials);
             return new AuthenticationResponseDTO()
             {
